Update PokeCull
This is probably my final version of the script. It's not perfected by any means. For some reason, the program prints blank lines equal to the number of times the user presses arrow keys, I couldn't get the Enter key to work correctly, and the tiny lag between screen refreshes is still present. However! It works lol.